correct "actions" description
During one of the exercises in the ETHZ-AMoD20 class we noticed that the docs are wrong here. From playing around with the simulator (on branch daffy), we noticed that the numbers in the "actions" variable don't correspond to forward velocity and steering angle but actually to the respective wheel inputs.